@@ -20,92 +20,98 @@
     "url": "git://github.com/veg/hyphy-vision.git"
   },
   "main": "dist/hyphyvision.js",
-  "version": "2.7.0",
+  "version": "2.8.0-alpha.1",
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.3.1",
-    "alignment.js": "0.1.1",
+    "@fortawesome/fontawesome-free": "^5.12.1",
+    "alignment.js": "0.3.0",
     "bootstrap": "4.x",
     "chi-squared": "^1.1.0",
+    "circos": "^2.1.0",
     "create-react-class": "^15.6.3",
-    "csvexport": "1.1.0",
+    "csvexport": "2.0.0",
     "d3": "3.x",
     "d3-save-svg": "0.0.2",
     "express": "^4.16.2",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-saver": "^1.3.3",
-    "immutable": "3.8.1",
+    "file-saver": "^2.0.2",
+    "immutable": "3.8.2",
     "in-browser-download": "^1.0.0",
     "jquery": "1.x",
     "jquery-ui": "1.x",
     "jquery-ui-bundle": "^1.11.4",
     "jquery-ui-dist": "^1.12.1",
-    "json-loader": "^0.5.6",
-    "lodash": "3.x",
+    "json-loader": "^0.5.7",
+    "lodash": "4.x",
     "lodash.clonedeep": "^4.5.0",
     "phylotree": "0.1.6",
     "popper.js": "^1.14.1",
     "pretty-data": "^0.40.0",
     "prop-types": "^15.5.10",
-    "react": "16",
+    "react": "16.12.0",
     "react-copy-to-clipboard": "^5.0.0",
-    "react-dom": "16",
+    "react-dom": "16.12.0",
     "react-json-view": "^1.14.0",
-    "react-phylotree": "0.1.0",
-    "react-router-dom": "^4.2.2",
+    "react-phylotree": "0.4.1",
+    "react-range": "^1.5.3",
+    "react-router-dom": "^5.1.2",
     "react-router-hash-link": "^1.1.1",
-    "react-scrollchor": "^4.2.1",
+    "react-scrollchor": "^6.0.0",
     "save-svg-as-png": "^1.2.0",
-    "supervisor": "^0.12.0",
-    "underscore": "^1.9.1"
+    "underscore": "1.9"
   },
   "devDependencies": {
-    "babel-core": "^6.17.0",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^19.0.0",
-    "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-1": "^6.24.1",
-    "base-href-webpack-plugin": "^1.0.2",
+    "@babel/core": "^7.8.7",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.8.7",
+    "@babel/preset-es2015": "^7.0.0-beta.53",
+    "@babel/preset-react": "^7.8.3",
+    "@fortawesome/fontawesome-svg-core": "^1.2.27",
+    "@fortawesome/free-brands-svg-icons": "^5.12.1",
+    "@fortawesome/free-regular-svg-icons": "^5.12.1",
+    "@fortawesome/free-solid-svg-icons": "^5.12.1",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.0.6",
+    "base-href-webpack-plugin": "^2.0.0",
     "bootstrap-webpack": "0.0.x",
-    "copy-webpack-plugin": "^4.1.1",
-    "css-loader": "^0.25.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^3.4.2",
     "electron": "^1.7.9",
     "enzyme": "3",
     "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^4.0.0",
-    "eslint-loader": "^1.8.0",
+    "eslint": "^6.8.0",
+    "eslint-loader": "^4.0.0",
     "eslint-plugin-react": "^7.1.0",
-    "exports-loader": "^0.6.3",
+    "exports-loader": "^0.7.0",
     "expose-loader": "^0.7.1",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-loader": "^0.9.0",
-    "font-awesome-webpack": "0.0.4",
-    "html-webpack-plugin": "^2.30.1",
-    "husky": "^0.14.3",
-    "imports-loader": "^0.6.5",
-    "jest": "^19.0.2",
-    "less": "^2.7.2",
-    "less-loader": "^4.0.4",
+    "file-loader": "^5.1.0",
+    "font-awesome-webpack": "^0.0.5-beta.2",
+    "font-awesome-webpack-4": "^1.0.0",
+    "html-webpack-plugin": "^4.0.0-alpha.2",
+    "husky": "^4.2.3",
+    "imports-loader": "^0.8.0",
+    "interpolate-html-plugin": "^3.0.0",
+    "jest": "^25.1.0",
+    "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.8.3",
     "npm-version": "^1.1.0",
-    "prettier": "1.18.1",
-    "pretty-quick": "^1.6.0",
+    "preload-webpack-plugin": "3.0.0-beta.3",
+    "prettier": "1.19.1",
+    "pretty-quick": "^2.0.1",
     "react-addons-test-utils": "^15.6.0",
     "react-lazy-render": "0.3.0",
     "react-test-renderer": "16",
-    "sass-loader": "^6.0.7",
+    "sass": "^1.26.2",
+    "sass-loader": "^8.0.2",
     "scss-loader": "^0.0.1",
-    "style-loader": "^0.13.1",
+    "style-loader": "^1.1.3",
     "uglify-js": "^3.3.7",
-    "url-loader": "^0.5.7",
-    "webpack": "^2.0.0",
-    "webpack-dev-server": "^2.9.7",
+    "url-loader": "^3.0.0",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
     "webpack-eslint-plugin": "^1.2.0"
   },
   "scripts": {
-    "build": "webpack",
+    "build": "webpack --mode production",
     "build:library": "webpack --config webpack.config.library.js",
     "lint": "eslint --ext .js,.jsx --ignore-pattern public,static,dist ./src/",
     "test": "node ./node_modules/jest-cli/bin/jest.js",
