@@ -1,8 +1,6 @@
-var React = require("react"),
-  ReactDOM = require("react-dom"),
-  _ = require("underscore");
+var React = require("react");
 
-const bomb = require("../../../images/bomb.svg");
+import bomb from "../../../images/bomb.svg";
 
 class ErrorBoundary extends React.Component {
   constructor(props) {
@@ -35,4 +33,4 @@ class ErrorBoundary extends React.Component {
   }
 }
 
-module.exports.ErrorBoundary = ErrorBoundary;
+export { ErrorBoundary };
