@@ -1,4 +1,4 @@
-var $ = require("jquery");
+const $ = require("jquery"); 
 
 var root = this;
 var datamonkey = function() {};
