@@ -24,11 +24,11 @@ function Header(props) {
   }
   $('[data-toggle="popover"]').popover();
   return (
-    <h4 className="dm-table-header">
+    <h4 className="dm-table-header mb-3">
       {props.title}
       {popover}
     </h4>
   );
 }
 
-module.exports.Header = Header;
+export { Header };
