@@ -34,7 +34,7 @@ function ScrollSpy(props) {
     );
   });
   return (
-    <nav className="col-sm-2 bs-docs-sidebar list-group d-xs-none d-sm-none d-md-none d-lg-none d-xl-block">
+    <nav className="col-sm-2 bs-docs-sidebar list-group d-none d-xs-none d-sm-none d-md-none d-lg-none d-xl-block">
       <ul className="nav nav-tabs nav-pills nav-stacked fixed flex-column">
         {list_items}
       </ul>
@@ -42,4 +42,4 @@ function ScrollSpy(props) {
   );
 }
 
-module.exports.ScrollSpy = ScrollSpy;
+export { ScrollSpy };
