@@ -1,11 +1,12 @@
 require("bootstrap");
 require("./datamonkey/datamonkey.js");
-require("./application.scss");
+//require("./application.scss");
 
 // Create new hyphy-vision export
 export { default as absrel } from "./jsx/absrel.jsx";
 export { default as busted } from "./jsx/busted.jsx";
 export { default as fel } from "./jsx/fel.jsx";
+export { default as ContrastFEL } from "./jsx/contrast-fel.jsx";
 export { default as meme } from "./jsx/meme.jsx";
 export { default as prime } from "./jsx/prime.jsx";
 export { default as relax } from "./jsx/relax.jsx";
@@ -15,11 +16,10 @@ export { default as gard } from "./jsx/gard.jsx";
 export { default as bgm } from "./jsx/bgm.jsx";
 export { default as fade } from "./jsx/fade.jsx";
 export { default as slatkin } from "./jsx/slatkin.jsx";
+export { default as MultiHit } from "./jsx/multihit.jsx";
 
 // renderHyPhyVision is for use in Galaxy
 export { default as renderHyPhyVision } from "./renderHyPhyVision.js";
 
 // Shared_component exports.
-export {
-  default as render_branch_selection
-} from "./jsx/shared_components/branch_selection.jsx";
+export { default as render_branch_selection } from "./jsx/shared_components/branch_selection.jsx";
